pyproject.toml
[tool.poetry]
name = "coc"
version = "0.1.0"
description = ""
authors = ["nanoguo <nanoguo@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
requests = "^2.28.1"
celery = "^5.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
